Add support for parsing BER
