+++ Sources/SwiftASN1/Basic ASN1 Types/UTCTime.swift
@@ -20,7 +20,7 @@
 /// ``UTCTime`` differs from ``GeneralizedTime`` in that it only has support for a two-digit year. This
 /// means that it can only encode dates between 1950 and 2049. For dates outside that range, prefer
 /// ``GeneralizedTime``.
-public struct UTCTime: DERImplicitlyTaggable, Hashable, Sendable {
+public struct UTCTime: DERImplicitlyTaggable, BERImplicitlyTaggable, Hashable, Sendable {
     @inlinable
     public static var defaultIdentifier: ASN1Identifier {
         .utcTime
@@ -137,6 +137,11 @@ public struct UTCTime: DERImplicitlyTaggable, Hashable, Sendable {
         }
 
         self = try TimeUtilities.utcTimeFromBytes(content)
+    }
+
+    @inlinable
+    public init(berEncoded node: ASN1Node, withIdentifier identifier: ASN1Identifier) throws {
+        self = try .init(derEncoded: node, withIdentifier: identifier)
     }
 
     @inlinable
